split page_layout.py into one file per element
